backend/requirements.txt
asgiref==3.8.1
astroid==3.3.10
dill==0.4.0
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
isort==6.0.1
mccabe==0.7.0
platformdirs==4.3.8
PyJWT==2.9.0
pylint==3.3.7
sqlparse==0.5.3
tomlkit==0.13.3
